added visual background, white card container and support link
