+++ src/components/AdvisorForm.tsx
@@ -1,11 +1,10 @@
-import { useState, useCallback, useEffect } from 'react';
+import { useState, useCallback } from 'react';
 import type { AdvisorFormState, FirmEntry, ContactFormData } from '../types';
 import { FirmService } from '../services/FirmService';
 import FirmInputStep from './FirmInputStep';
 import ContactDetailsStep from './ContactDetailsStep';
 import FormCompleteStep from './FormCompleteStep';
 import ErrorBoundary from './ErrorBoundary';
-import Toast from './Toast';
 
 const initialFormState: AdvisorFormState = {
   currentStep: 'firm-input',
@@ -19,16 +18,15 @@ const initialFormState: AdvisorFormState = {
 
 interface AdvisorFormProps {
   onComplete?: (firms: FirmEntry[], userEmail: string) => void;
+  onToast?: (message: string) => void;
 }
 
-export function AdvisorForm({ onComplete }: AdvisorFormProps) {
+export function AdvisorForm({ onComplete, onToast }: AdvisorFormProps) {
   const [formState, setFormState] = useState<AdvisorFormState>(initialFormState);
   const [loading, setLoading] = useState(false);
   const [firmInputError, setFirmInputError] = useState<string>('');
   const [currentFirmInput, setCurrentFirmInput] = useState<string>('');
   const [emailError, setEmailError] = useState<string>('');
-  const [toastMessage, setToastMessage] = useState<string>('');
-  const [showToast, setShowToast] = useState(false);
 
   const remainingFirms = formState.maxFirms - formState.enteredFirms.length;
 
@@ -89,8 +87,9 @@ export function AdvisorForm({ onComplete }: AdvisorFormProps) {
       }));
 
       // Show toast notification
-      setToastMessage(`Thank you! We'll be in touch with you about ${exactFirmName}.`);
-      setShowToast(true);
+      if (onToast) {
+        onToast(`Thank you! We'll be in touch with you about ${exactFirmName}.`);
+      }
     }
   }, [formState.enteredFirms, formState.userEmail]);
 
@@ -122,8 +121,9 @@ export function AdvisorForm({ onComplete }: AdvisorFormProps) {
       }));
 
       // Show toast notification for matched firm
-      setToastMessage(`Thank you! We'll be in touch with you about ${formState.currentFirmName}.`);
-      setShowToast(true);
+      if (onToast) {
+        onToast(`Thank you! We'll be in touch with you about ${formState.currentFirmName}.`);
+      }
 
       console.log('Returning to firm-input step');
       setLoading(false);
@@ -139,16 +139,6 @@ export function AdvisorForm({ onComplete }: AdvisorFormProps) {
     }));
   }, []);
 
-  // Auto-dismiss toast after 3 seconds
-  useEffect(() => {
-    if (showToast) {
-      const timer = setTimeout(() => {
-        setShowToast(false);
-      }, 3000);
-
-      return () => clearTimeout(timer);
-    }
-  }, [showToast]);
 
   const handleFinish = useCallback(async () => {
     // Validate email is provided and valid
@@ -289,17 +279,6 @@ export function AdvisorForm({ onComplete }: AdvisorFormProps) {
           />
         )}
       </ErrorBoundary>
-
-      {/* Toast notification */}
-      {showToast && (
-        <div className="flex justify-center pt-4">
-          <Toast
-            label={toastMessage}
-            intent="success"
-            onClose={() => setShowToast(false)}
-          />
-        </div>
-      )}
     </div>
   );
 }
